+++ app/layouts/default.vue
@@ -10,8 +10,8 @@ const { data: navigation } = await useNavigation()
   <NavigationAnnouncementBanner v-if="navigation!.announcementBannerShow" v-bind="navigation!" />
   <NavigationHeader :dark-header />
   <slot />
-  <div fixed bottom-32 right-32 z-102>
-    <LockBadge :draft with-env />
+  <div fixed bottom-20 right-20 z-102>
+    <PageInfo :draft with-env />
   </div>
   <HexagonsBackground z-1 :bg-color="footerBgColor" :with-socials="showSocialsHexagonBg" />
   <NavigationFooter :bg-color="footerBgColor" />
